split basic usage into mapping and suggesting
